Moved Dashboards gets to own service
